Adding ObjectCube requirements to setup.py
We read the requirements.txt file and add it to the setup file.
If this module is now installed into a given virtual environment,
dependent packages are install as well.

Removed dependencies that are not required by ObjectCube. These are
development specific and developers can simply install them into their
own environment when needed.